Reduce warnings

+++ Src/Common/unicoder.cpp
@@ -482,7 +482,7 @@ bool maketstring(String & str, const char* lpd, size_t len, int codepage, bool *
 	LPWSTR wbuff = &*str.begin();
 	if (codepage == CP_ACP || IsValidCodePage(codepage))
 	{
-		int n = MultiByteToWideChar(codepage, flags, lpd, len, wbuff, wlen - 1);
+		int n = MultiByteToWideChar(codepage, flags, lpd, static_cast<int>(len), wbuff, static_cast<int>(wlen - 1));
 		if (n)
 		{
 			/*
@@ -516,7 +516,7 @@ bool maketstring(String & str, const char* lpd, size_t len, int codepage, bool *
 		{
 			if (GetLastError() == ERROR_INVALID_FLAGS)
 			{
-				int n = MultiByteToWideChar(codepage, 0, lpd, len, wbuff, wlen-1);
+				int n = MultiByteToWideChar(codepage, 0, lpd, static_cast<int>(len), wbuff, static_cast<int>(wlen-1));
 				if (n)
 				{
 					/* NB: MultiByteToWideChar is documented as only zero-terminating 
@@ -547,7 +547,7 @@ bool maketstring(String & str, const char* lpd, size_t len, int codepage, bool *
 				*lossy = true;
 				flags = 0;
 				// wlen & wbuff are still fine
-				n = MultiByteToWideChar(codepage, flags, lpd, len, wbuff, wlen-1);
+				n = MultiByteToWideChar(codepage, flags, lpd, static_cast<int>(len), wbuff, static_cast<int>(wlen-1));
 				if (n)
 				{
 					try
@@ -691,14 +691,14 @@ int CrossConvert(const char* src, unsigned srclen, char* dest, unsigned destsize
 	if (cpin == CP_UCS2LE)
 	{
 		if (srclen == -1)
-			srclen = wcslen((wchar_t *)src) * sizeof(wchar_t);
+			srclen = static_cast<unsigned>(wcslen((wchar_t *)src) * sizeof(wchar_t));
 		memcpy(wbuff.get(), src, srclen);
 		n = srclen / sizeof(wchar_t);
 	}
 	else if (cpin == CP_UCS2BE)
 	{
 		if (srclen == -1)
-			srclen = wcslen((wchar_t *)src) * sizeof(wchar_t);
+			srclen = static_cast<unsigned>(wcslen((wchar_t *)src) * sizeof(wchar_t));
 		_swab((char *)src, (char *)wbuff.get(), srclen);
 		n = srclen / sizeof(wchar_t);
 	}
@@ -999,10 +999,10 @@ bool convert(UNICODESET unicoding1, int codepage1, const unsigned char * src, si
 		if (destcp == CP_ACP || IsValidCodePage(destcp))
 		{
 			DWORD flags = 0;
-			int bytes = WideCharToMultiByte(destcp, flags, (LPCWSTR)src, srcbytes/2, 0, 0, NULL, NULL);
+			int bytes = WideCharToMultiByte(destcp, flags, (LPCWSTR)src, static_cast<int>(srcbytes/2), 0, 0, NULL, NULL);
 			dest->resize(bytes + 2);
 			int losses = 0;
-			bytes = WideCharToMultiByte(destcp, flags, (LPCWSTR)src, srcbytes/2, (char *)dest->ptr, dest->capacity, NULL, NULL);
+			bytes = WideCharToMultiByte(destcp, flags, (LPCWSTR)src, static_cast<int>(srcbytes/2), (char *)dest->ptr, static_cast<int>(dest->capacity), NULL, NULL);
 			dest->ptr[bytes] = 0;
 			dest->ptr[bytes+1] = 0;
 			dest->size = bytes;
@@ -1033,9 +1033,9 @@ bool convert(UNICODESET unicoding1, int codepage1, const unsigned char * src, si
 		if (srccp == CP_ACP || IsValidCodePage(srccp))
 		{
 			DWORD flags = 0;
-			int wchars = MultiByteToWideChar(srccp, flags, (LPCSTR)src, srcbytes, 0, 0);
+			int wchars = MultiByteToWideChar(srccp, flags, (LPCSTR)src, static_cast<int>(srcbytes), 0, 0);
 			dest->resize((wchars + 1) *2);
-			wchars = MultiByteToWideChar(srccp, flags, (LPCSTR)src, srcbytes, (LPWSTR)dest->ptr, dest->capacity/2);
+			wchars = MultiByteToWideChar(srccp, flags, (LPCSTR)src, static_cast<int>(srcbytes), (LPWSTR)dest->ptr, static_cast<int>(dest->capacity/2));
 			dest->ptr[wchars * 2] = 0;
 			dest->ptr[wchars * 2 + 1] = 0;
 			dest->size = wchars * 2;
@@ -1069,11 +1069,11 @@ bool convert(UNICODESET unicoding1, int codepage1, const unsigned char * src, si
  */
 static void convert(const std::wstring& from, unsigned codepage, std::string& to)
 {
-	int len = WideCharToMultiByte(codepage, 0, from.c_str(), from.length(), 0, 0, 0, 0);
+	int len = WideCharToMultiByte(codepage, 0, from.c_str(), static_cast<int>(from.length()), 0, 0, 0, 0);
 	if (len)
 	{
 		to.resize(len);
-		WideCharToMultiByte(codepage, 0, from.c_str(), from.length(), &to[0], len, NULL, NULL);
+		WideCharToMultiByte(codepage, 0, from.c_str(), static_cast<int>(from.length()), &to[0], static_cast<int>(len), NULL, NULL);
 	}
 	else
 	{
